Now handles user errors. With nice feedback to swipe right.

+++ Discount Calculator/Discount Calculator/CalculatorData.h
@@ -10,14 +10,14 @@
 
 @interface CalculatorData : NSObject
 @property float price;
-@property int dollarsOff;
-@property int discount;
-@property int additionalDiscount;
+@property float dollarsOff;
+@property float discount;
+@property float additionalDiscount;
 @property float tax;
 @property (readonly) float discountedPrice;
 @property (readonly) float originalPrice;
 
--(id) initWithData:(float) _price dollarsOff:(int) _dollarsOff discount:(int) _discount additionalDiscount:(int) _additionalDiscount tax:(float) _tax;
+-(id) initWithData:(float) _price dollarsOff:(float) _dollarsOff discount:(float) _discount additionalDiscount:(float) _additionalDiscount tax:(float) _tax;
 -(float) discountedPrice;
 -(float) originalPrice;
 +(CalculatorData*) mainPrice;
